requirements.txt
pyspark==3.5.0
requests
psycopg2-binary
